requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
pydantic[email]==2.4.2
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.3
pytest==7.4.3
pytest-asyncio==0.21.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
httpx==0.25.2
email-validator==2.1.0
ruff==0.1.6
black==23.12.1
pre-commit==3.5.0